[benchmark] report: Add MetricFillLatency and MetricFillThroughput
Extract from disk.c the logic of these new utilities.
